scripts/evaluate.py
#!/usr/bin/env python3
"""
Evaluate Trinity Cognitive Probes on real LLMs
Measuring Progress Toward AGI Hackathon 2026
"""

import json
import os
import sys
from pathlib import Path
from datetime import datetime
from typing import Dict, List, Optional
from dataclasses import dataclass, asdict


@dataclass
class Question:
    """Single MC question from Trinity Cognitive Probes"""
    id: str
    question_type: str
    question: str
    choices: List[str]
    answer: str


@dataclass
class EvaluationResult:
    """Result of evaluating a single question"""
    question_id: str
    track: str
    predicted: str
    correct: bool
    confidence: float
    reasoning: str


class ModelEvaluator:
    """Base class for evaluating models"""

    def __init__(self, model_name: str):
        self.model_name = model_name
        self.results: List[EvaluationResult] = []

    def evaluate(self, questions: List[Question]) -> List[EvaluationResult]:
        """Evaluate a list of questions"""
        raise NotImplementedError("Subclasses must implement evaluate()")

    def save_results(self, output_dir: str):
        """Save evaluation results to JSON"""
        output_path = Path(output_dir) / f"{self.model_name}_results.json"
        output_path.parent.mkdir(parents=True, exist_ok=True)

        with open(output_path, 'w') as f:
            json.dump([asdict(r) for r in self.results], f, indent=2)

        print(f"Results saved to {output_path}")
        self.print_summary()

    def print_summary(self):
        """Print evaluation summary"""
        total = len(self.results)
        correct = sum(1 for r in self.results if r.correct)

        print(f"\n{'='*60}")
        print(f"Evaluation Summary: {self.model_name}")
        print(f"{'='*60}")
        print(f"Total Questions: {total}")
        print(f"Correct: {correct}")
        print(f"Accuracy: {correct/total:.2%}")


class ClaudeEvaluator(ModelEvaluator):
    """Evaluate using Claude via API"""

    def __init__(self, model_name: str = "claude"):
        super().__init__(model_name)

    def evaluate(self, questions: List[Question]) -> List[EvaluationResult]:
        """Evaluate questions using Claude"""
        import anthropic

        # TODO: Implement Claude API evaluation
        # This will require setting up Anthropic API credentials
        print(f"Would evaluate {len(questions)} questions with Claude API")
        return []


class GeminiEvaluator(ModelEvaluator):
    """Evaluate using Google Gemini"""

    def __init__(self, model_name: str = "gemini"):
        super().__init__(model_name)

    def evaluate(self, questions: List[Question]) -> List[EvaluationResult]:
        """Evaluate questions using Gemini"""
        import google.generativeai as genai

        # TODO: Implement Gemini API evaluation
        print(f"Would evaluate {len(questions)} questions with Gemini API")
        return []


class OpenAIEvaluator(ModelEvaluator):
    """Evaluate using OpenAI models (GPT-4, etc.)"""

    def __init__(self, model_name: str = "openai"):
        super().__init__(model_name)

    def evaluate(self, questions: List[Question]) -> List[EvaluationResult]:
        """Evaluate questions using OpenAI API"""
        import openai

        # TODO: Implement OpenAI API evaluation
        print(f"Would evaluate {len(questions)} questions with OpenAI API")
        return []


def load_questions(track_dir: str) -> List[Question]:
    """Load MC questions from CSV file"""
    import csv

    questions = []
    csv_path = Path(track_dir)

    # Find CSV file in directory
    csv_files = list(csv_path.glob("*.csv"))
    if not csv_files:
        print(f"No CSV files found in {csv_path}")
        return []

    csv_file = csv_files[0]

    with open(csv_file, 'r') as f:
        reader = csv.DictReader(f)

        for row in reader:
            # Map CSV columns to Question dataclass
            questions.append(Question(
                id=row.get('id', ''),
                question_type=row.get('question_type', ''),
                question=row.get('question', ''),
                choices=[row.get('A', ''), row.get('B', ''), row.get('C', ''), row.get('D', '')],
                answer=row.get('answer', '')
            ))

    print(f"Loaded {len(questions)} questions from {csv_file}")
    return questions


def main():
    """Main evaluation script"""
    if len(sys.argv) < 2:
        print("Usage: python evaluate.py --model <claude|gemini|openai> --track <all|thlp|ttm|tagp|tefb|tscp>")
        sys.exit(1)

    model = None
    track = "all"

    args = sys.argv[1:]
    i = 0
    while i < len(args):
        if args[i] == "--model":
            model = args[i + 1]
            i += 2
        elif args[i] == "--track":
            track = args[i + 1]
            i += 2
        else:
            i += 1

    if not model:
        print("Error: --model is required")
        print("Options: claude, gemini, openai")
        sys.exit(1)

    # Create evaluator based on model
    evaluators = {
        'claude': ClaudeEvaluator,
        'gemini': GeminiEvaluator,
        'openai': OpenAIEvaluator
    }

    evaluator = evaluators[model]()

    # Load questions
    base_dir = Path(__file__).parent.parent / "data"

    tracks_to_run = []
    if track == "all":
        tracks_to_run = ["thlp", "ttm", "tagp", "tefb", "tscp"]
    else:
        tracks_to_run = [track]

    print(f"\n{'='*60}")
    print(f"Running evaluations for: {evaluator.model_name}")
    print(f"Tracks: {', '.join(tracks_to_run)}")
    print(f"{'='*60}\n")

    all_results = []

    for t in tracks_to_run:
        track_dir = base_dir / t
        if not track_dir.exists():
            print(f"Warning: Track directory {track_dir} not found, skipping...")
            continue

        print(f"\nEvaluating track: {t}")
        questions = load_questions(str(track_dir))

        if questions:
            results = evaluator.evaluate(questions)
            evaluator.results.extend(results)

    # Save results
    if evaluator.results:
        evaluator.save_results("runs")


if __name__ == "__main__":
    main()
