requirements.txt
# Dependencies for AGI Hackathon Evaluation

# API Clients
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
